UGENE-4738 fix: Arguments in all qMin() and qMax() functions must have the same type
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@11183 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Algorithm/src/util_orf/ORFFinder.cpp
@@ -276,7 +276,7 @@ void ORFFindAlgorithm::find(
                 if((seqPointer % BLOCK_READ_FROM_DB) == 0){// query to db
                     sequence.clear();
                     QByteArray tmp;
-                    qint64 regStart = qMax((qint64)maxInitiator, (qint64)i - (BLOCK_READ_FROM_DB + 3));
+                    qint64 regStart = qMax((qint64)maxInitiator, i - (BLOCK_READ_FROM_DB + 3));
                     qint64 regLen = qMin(i - maxInitiator + 1, (qint64)BLOCK_READ_FROM_DB + 3 + 1);
                     tmp.append(dnaSeq.getSequenceData(U2Region(regStart, regLen), os));
                     SAFE_POINT_OP(os, );
