+++ src/plugins/external_tool_support/src/ceas/CEASSupportTask.cpp
@@ -256,7 +256,7 @@ int CEASLogParser::getProgress(){
         if(lastMessage.contains(rx)){
             SAFE_POINT(rx.indexIn(lastMessage) > -1, "bad progress index", 0);
             int step = rx.cap(1).toInt();
-            return  (100 * step)/ float(qMax(step, max_step));
+            return static_cast<int>((100 * step)/ float(qMax(step, max_step)));
         }
     }
     return progress;
